[CALCITE-4963] Review comments
