+++ core/src/main/java/org/apache/calcite/sql/SqlDialectFactoryImpl.java
@@ -16,8 +16,6 @@
  */
 package org.apache.calcite.sql;
 
-import org.apache.calcite.avatica.util.Casing;
-import org.apache.calcite.config.NullCollation;
 import org.apache.calcite.sql.dialect.AccessSqlDialect;
 import org.apache.calcite.sql.dialect.AnsiSqlDialect;
 import org.apache.calcite.sql.dialect.BigQuerySqlDialect;
@@ -54,7 +52,6 @@ import org.apache.calcite.sql.dialect.VerticaSqlDialect;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.SQLException;
 import java.util.Locale;
@@ -69,27 +66,8 @@ public class SqlDialectFactoryImpl implements SqlDialectFactory {
       JethroDataSqlDialect.createCache();
 
   @Override public SqlDialect create(DatabaseMetaData databaseMetaData) {
-    SqlDialect.Context context = Utils.context(databaseMetaData);
-    return create(databaseMetaData, context, new AnsiSqlDialect(context));
-  }
+    SqlDialect.Context context = SqlDialects.createContext(databaseMetaData);
 
-  /**
-   * Creates a <code>SqlDialect</code>.
-   *
-   * <p>Does not maintain a reference to the DatabaseMetaData -- or, more
-   * importantly, to its {@link Connection} -- after this call has
-   * returned.
-   *
-   * @param databaseMetaData used to determine which dialect of SQL to
-   *                         generate
-   * @param context          the pre-built {@link SqlDialect.Context}
-   * @param fallback         the default {@link SqlDialect} in case none of the selection rules match
-   *
-   * @throws RuntimeException if there was an error creating the dialect
-   */
-  public SqlDialect create(DatabaseMetaData databaseMetaData,
-                           SqlDialect.Context context,
-                           SqlDialect fallback) {
     String databaseProductName = context.databaseProductName();
     try {
       if (databaseProductName == null) {
@@ -98,7 +76,6 @@ public class SqlDialectFactoryImpl implements SqlDialectFactory {
     } catch (SQLException e) {
       throw new RuntimeException("while detecting database product", e);
     }
-
     String upperProductName = databaseProductName.toUpperCase(Locale.ROOT).trim();
 
     switch (upperProductName) {
@@ -175,74 +152,7 @@ public class SqlDialectFactoryImpl implements SqlDialectFactory {
     } else if (upperProductName.contains("SPARK")) {
       return new SparkSqlDialect(context);
     } else {
-      return fallback;
-    }
-  }
-
-  private static Casing getCasing(DatabaseMetaData databaseMetaData, boolean quoted) {
-    try {
-      if (quoted
-          ? databaseMetaData.storesUpperCaseQuotedIdentifiers()
-          : databaseMetaData.storesUpperCaseIdentifiers()) {
-        return Casing.TO_UPPER;
-      } else if (quoted
-          ? databaseMetaData.storesLowerCaseQuotedIdentifiers()
-          : databaseMetaData.storesLowerCaseIdentifiers()) {
-        return Casing.TO_LOWER;
-      } else if (quoted
-          ? (databaseMetaData.storesMixedCaseQuotedIdentifiers()
-              || databaseMetaData.supportsMixedCaseQuotedIdentifiers())
-          : (databaseMetaData.storesMixedCaseIdentifiers()
-              || databaseMetaData.supportsMixedCaseIdentifiers())) {
-        return Casing.UNCHANGED;
-      } else {
-        return Casing.UNCHANGED;
-      }
-    } catch (SQLException e) {
-      throw new IllegalArgumentException("cannot deduce casing", e);
-    }
-  }
-
-  private static boolean isCaseSensitive(DatabaseMetaData databaseMetaData) {
-    try {
-      return databaseMetaData.supportsMixedCaseIdentifiers()
-          || databaseMetaData.supportsMixedCaseQuotedIdentifiers();
-    } catch (SQLException e) {
-      throw new IllegalArgumentException("cannot deduce case-sensitivity", e);
-    }
-  }
-
-  private static NullCollation getNullCollation(DatabaseMetaData databaseMetaData) {
-    try {
-      if (databaseMetaData.nullsAreSortedAtEnd()) {
-        return NullCollation.LAST;
-      } else if (databaseMetaData.nullsAreSortedAtStart()) {
-        return NullCollation.FIRST;
-      } else if (databaseMetaData.nullsAreSortedLow()) {
-        return NullCollation.LOW;
-      } else if (databaseMetaData.nullsAreSortedHigh()) {
-        return NullCollation.HIGH;
-      } else if (isBigQuery(databaseMetaData)) {
-        return NullCollation.LOW;
-      } else {
-        throw new IllegalArgumentException("cannot deduce null collation");
-      }
-    } catch (SQLException e) {
-      throw new IllegalArgumentException("cannot deduce null collation", e);
-    }
-  }
-
-  private static boolean isBigQuery(DatabaseMetaData databaseMetaData)
-      throws SQLException {
-    return databaseMetaData.getDatabaseProductName()
-        .equals("Google Big Query");
-  }
-
-  private static String getIdentifierQuoteString(DatabaseMetaData databaseMetaData) {
-    try {
-      return databaseMetaData.getIdentifierQuoteString();
-    } catch (SQLException e) {
-      throw new IllegalArgumentException("cannot deduce identifier quote string", e);
+      return new AnsiSqlDialect(context);
     }
   }
 
@@ -315,52 +225,6 @@ public class SqlDialectFactoryImpl implements SqlDialectFactory {
     case UNKNOWN:
     default:
       return null;
-    }
-  }
-
-  /**
-   * Utilities for {@link SqlDialectFactory} to extract information from {@link DatabaseMetaData}.
-   */
-  public static final class Utils {
-    private Utils() {
-      // utility class - not to be instantiated
-    }
-
-    /**
-     * Extracts information from {@link DatabaseMetaData} into {@link SqlDialect.Context}.
-     * @param databaseMetaData the database metadata
-     * @return a context with information populated from the database metadata
-     */
-    public static SqlDialect.Context context(DatabaseMetaData databaseMetaData) {
-      String databaseProductName;
-      int databaseMajorVersion;
-      int databaseMinorVersion;
-      String databaseVersion;
-      try {
-        databaseProductName = databaseMetaData.getDatabaseProductName();
-        databaseMajorVersion = databaseMetaData.getDatabaseMajorVersion();
-        databaseMinorVersion = databaseMetaData.getDatabaseMinorVersion();
-        databaseVersion = databaseMetaData.getDatabaseProductVersion();
-      } catch (SQLException e) {
-        throw new RuntimeException("while detecting database product", e);
-      }
-      final String quoteString = getIdentifierQuoteString(databaseMetaData);
-      final NullCollation nullCollation = getNullCollation(databaseMetaData);
-      final Casing unquotedCasing = getCasing(databaseMetaData, false);
-      final Casing quotedCasing = getCasing(databaseMetaData, true);
-      final boolean caseSensitive = isCaseSensitive(databaseMetaData);
-      final SqlDialect.Context c = SqlDialect.EMPTY_CONTEXT
-          .withDatabaseProductName(databaseProductName)
-          .withDatabaseMajorVersion(databaseMajorVersion)
-          .withDatabaseMinorVersion(databaseMinorVersion)
-          .withDatabaseVersion(databaseVersion)
-          .withIdentifierQuoteString(quoteString)
-          .withUnquotedCasing(unquotedCasing)
-          .withQuotedCasing(quotedCasing)
-          .withCaseSensitive(caseSensitive)
-          .withNullCollation(nullCollation);
-
-      return c;
     }
   }
 }
